Invalidate metadata if errors were previously ignored (#3793)
Using --incremental with --follow-imports=silent triggered a bug where a
file was not typechecked due to the presence of a metadata file
and accompanying AST from a previous run and would therefore not print any
errors. This diff fixes the issue by recording whether errors were
silenced in the metadata file when it was created and ignoring these files if
errors are no longer silenced.