Encapsulate callable name access in messages.py (#4136)
Trying to solve #4133 in two steps, in order to avoid conflicts on semanal.py.
This step does not do much by itself - it replaces callee.name with
callable_name(callee). In later steps, the quotation marks can be added inside
callable_name (which is a formatting function) instead of being part of
CallableType.