Modifying multiple files per update in fine-grained incremental mode (#4282)
The approach is to split multiple file updates into multiple single-file
updates, and only report errors for the final update. If a modified file
brings in additional files through an import, we'll process one of the
imported files instead, changing the target file as we go. The
motivation is that this should result in less redundant work than
postponing the imported file, as a missing imported file will result
in a lot of stale targets which would need to processed again on the
next update.

This still doesn't handle blockers well. I'll fix these issues separately.

This is inefficient as some files can be parsed multiple times, with
O(n**2) worst case behavior (n is the number of changed files). It
should be fine when a small number of files need to be processed.