Clean up decorator implementation (#4325)
Store decorated function as Decorator in symbol table, not as Var.
Previously module-level function were stored as Var nodes in symbol
tables. That was inconsistent, as decorated methods were stored as
Decorators in symbol tables.

As a side effect, this also fixes some cases of conditional method
definition that were incorrectly rejected previously.

This helps with fine-grained incremental checking, as it gives us
access to a decorated function through the module symbol table.