Fine-grained: Don't crash if passed a non-existent file to dmypy (#4471)
This fixes crashes, but the current behavior is still not correct, as
non-existent files are not reported as errors in fine-grained
incremental updates. Always detecting these is somewhat non-trivial,
since currently `mypy.server.update` doesn't get information about which
files were explicitly passed as arguments -- it's okay to files to be
deleted, as long they weren't explicitly passed on the command
line. I'm not fixing this issue here since it's relatively benign,
whereas the crashes are blocking progress.

Also update the test runner to support passing non-existent paths as
arguments. These need to be tagged using magic module name
prefixes/suffixes, as we normally check that there are no misspelled
module names in test case descriptions by checking that the modules
passed on the command line exist.