Fine-grained: Treat files changed only if md5 changes (#4505)
Previously the daemon could do a lot of extra work if you switched
to another git branch and immediately back to the original branch,
since many file timestamps would change. Now this isn't sufficient
to consider a file as changed.

Also add a new cached file system abstraction (that is currently only
used in dmypy) to avoid redundant file system operations and to make
file system state easier to reason about. The idea is that we cache
output of previous file system operations during a single increment,
for both consistency and performance. My plan is to eventually use it
about everywhere.

This should also make it slightly easier to switch to listening to
file system events instead of stat()ing everything, but that's not
a priority yet.

Fixes #4499.