Optionally check that we don't have duplicate nodes after AST merge (#4647)
It's hard to manually ensure that all relevant nodes are merged in AST merge
and that no references to stale objects remain. This automated opt-in check
looks for these duplicate nodes. It can be enabled in tests by switching
`mypy.test.testfinegrained.CHECK_CONSISTENCY` to True.

This was originally implemented by @msullivan. I did various minor changes
and some refactoring.