Check Liskov compatibility in fine-grained mode (#4961)
Fixes #4960
Fixes #4853

The first issue is fixed by:

* Defining a new Var if there is a self assignment with an explicit 
  type in a subclass.
* Applying check_compatibility_super to RefExpr lvalues, not only 
  NameExpr.

The second issue is fixed by adding two dependencies:

* If a variable is defined in class, recheck the current target (without 
  recursing to methods).
* If a variable is defined via self in a method, recheck the defining 
  method only.
* The above two are only added for classes with user-defined bases 
  (this allows to avoid trivial deps that would only "pollute" many tests, 
  but are useless in practice).

There was one more thing I tried to fix, there seems to be an 
inconsistency that lvalues corresponding to self definitions have their 
kind set to None instead of MDEF (like for names defined in class). But 
I abandoned this because some code relies on this. I think we can clean 
this in an upcoming refactoring.