Fix fine-grained crash caused by unneeded extra parse_file()s (#4980)
load_graph() contains code to parse ancestor modules that
get new children. Since load_graph() did not communicate that
these modules had been parsed to fine-grained mode, this caused
problems with fine-grained mode.

That code was added in #1865 to fix a bug where 'import from'
of a newly added submodule would not work in incremental mode
because it didn't appear in the parent module's symbol table.
The need to reload the parent to fix this was obviated by
at the modules map to find possible submodules.

Since the code isn't needed any more and it is causing trouble,
we just remove it.