Support protocols in fine grained mode (#4790)
Support of protocols in fine grained mode is non-trivial because protocols are not very "modular" (can have action at a distance). To support them we introduce several new dependencies and a global state that represents protocol dependency map (serialized in a separate cache file).

There are three kinds of protocol dependencies. For example, after a subtype check:
x: Proto = C()
the following dependencies will be generated:
1. ..., <SuperProto[wildcard]>, <Proto[wildcard]> -> <Proto>
2. ..., <B.attr>, <C.attr> -> <C> [for every attr in Proto members]
3. <C> -> Proto  # this one to invalidate the subtype cache
The first kind is generated immediately per-module. While two other kinds are generated
after all modules are type checked and we have recorded all the subtype checks.

Another change is that we reset subtype caches in protocols scheduled for recheck _before_ processing any other targets. This avoids some false negatives in corner cases that are sensitive to module processing order.

I didn't test the scenario when the protocol cache files are tampered with, it is also hard to write any tests for this.