Make coroutine function return type more specific (#5052)
* Change the return type of coroutine functions to coroutine object

Previously, the return type was `Awaitable`, which is correct, but
not specific enough for some use cases.  For example, if you have
a function parameter that should be a coroutine object (but not a
`Future`, `Task` or other awaitable), then `mypy` is unable to
detect incorrect invocations of the function.

This change is (deliberately) imcomplete as is.  It seems like
this breaks quite a few tests in `mypy`.  The first symptom is that
coroutine objects are now (incorrectly) detected as generators..

* Prevent coroutine functions from being classified as generators

This change removes the undesired "return type of a generator
function should be `Generator` or one of its subtypes for coroutine
function definitions.

However, it introduces a new error where the return type of coroutine
functions is expected to be `Coroutine[T, Any, Any]` instead of the
desired `T`.  It looks like we were hijacking a generator-specific
path for checking the return type of coroutine functinos.  I added
an explicit path for coroutine functions, allowing our test to pass.

However, lots of tests now fail.  A few of them were simply places
that were incidentally relying on coroutine functions to have type
`Awaitable`.  I fix them.  The remaining failures all seem to be
about coroutine functions with return type `None` without an explicit
return statement.  Seems like this is also something for which we
were relying on implicit classification as generators.

* Allow implicit return for coroutine functions that return `None`

Most of the tests are fixed, but two tests still fail.  One about
not detecting invalid `yield from` on `AwaitableGenerator`.  The
other about types being erased in call to `asyncio.gather()`.

* Fix return type for coroutine functions decorated with @coroutine

* Fix detection of await expression on direct coroutine function call

Changing the return type of coroutine functions to  `Coroutine`
introduced a regression in the expression checks.

* Fix regression after change of coroutine function return type

* Fix position of return type in `Coroutine`

This fixes the type inference logic that was causing the last
failing test to fail.  Build should now be green :-)

* Fix issues raised in code review

Fixes #3569.
Fixes #4460.

Special thanks to @ilevkivskyi and @gvanrossum for their time
and their infinite patience with all my questions :-)