Basic support for user-defined generics  (mypyc/mypyc#116)
Support defining simple generic functions and classes. Support type
applications.

The idea is to erase type variable types to the upper bound, and to
use the declared function signatures and attribute types when
generating coercions instead of the inferred types, since the latter
are generally too precise. For the rest we just rely on the existing
machinery to add implicit coercions.