Rework vtable handling, emit glue methods when override changes method type (mypyc/mypyc#154)
When we have:
```    
class A:
    def f(self, x: int) -> object: ...
class B(A):
    def f(self, x: object) -> int: ...
```
then `A.f` and `B.f` have different function signatures in the generated C, 
because `int` and `object` have different runtime representations.

So `B` then needs to have two versions of `f` in its symbol table: one for
when it is called as an `A`, and one for as a `B`. This means we need to
generate a glue function to convert from `A.f` to `B.f`.

To support this, we restructure a lot of how vtables are represented in the compiler.