Require abstract classes in stubs to be explicitly marked (#5223)
Fixes #3692.

In typeshed, we have had many bugs because of classes that were not meant to be abstract, but were inferred by mypy to be abstract because they didn't implement some methods from an abstract base class. I call such classes "accidentally abstract"; python/typeshed#1476 has a list of cases where this bug appeared.

It is hard to catch these bugs because they only show up in user code when somebody tries to instantiate the class, and when reviewing changes to typeshed you have to manually go over all methods to find instances of this bug. Therefore, I'd like to have mypy point out accidentally abstract classes automatically.

This PR throws an error if a class:
- Has any abstract method;
- Does not itself have any method decorated with `@abstractmethod` (i.e., all abstract methods are inherited);
- Does not have `ABCMeta` as an explicit metaclass, or `Protocol` as an explicit base class;
- Is defined in a stub file.