Support import cycles (mypyc/mypyc#195)
This approach is pretty janky -- the module init function can be called
multiple times if there are import cycles, and we just return the previously
created module object (which may be only partially initialized) if it's 
available. This Seems To Work but might break in some unforeseen ways.
However, I propose that we keep this for now, at least until we support
module top levels -- as that's where the potential problems with this 
approach seem likely to materialize.

I manually verified that references to other modules within an import
cycle seem to generate efficient code (i.e. not go through Python 
semantics) but there are no tests for that yet. I'll create a follow-up
issue about this.

Fixes mypyc/mypyc#164.