Allow __getattr__ during semantic analysis (#5295)
Previously module `__getattr__` was used during type checking (runtime context). But it was only partially allowed in semantic analysis (after `from a import A`, `A` can be used in annotations/base classes). This PR allows the same support for `a.A`, i.e. it can be used in annotations and base classes (provided `__getattr__` has a suitable return type `Any`).