Finish implementation of try/catch (mypyc/mypyc#251)
This adds support for multiple except blocks with exception types that
are matched against and binding the exception to a variable.

It also fixes handling of sys.exc_info(): when we finish handling an
exception (either naturally or because an exception is raised), we
need to restore the previous exc_info. To manage this we have
CPy_CatchError return the previous exc_info, which is then restored
when leaving the except block. Access to the exception being handled
is then done via exc_info, where it needs to be anyways.