Semantically analyze indexes in type var definitions (#5376)
Support type aliases in mypyc is nearly a matter of just directly
compiling the assignment expressions. The snag is that mypy doesn't
semantically analyze the indexes in index expressions when defining
type variables (it instead just does type things).

Update semanal to semantically analyze the indexes. I just do this
unconditionally because it seems cheap enough that it doesn't seem
worth introducing mechanism for doing it conditionally.