Implement a restricted form of attribute default values (mypyc/mypyc#331)
The necessity of interoperating properly via the C API makes it
difficult for us to faithfully implement class variables.

By default CPython relies on the interaction of instance `__dict__`s
and class `__dict__`s to provide the "default" behavior of falling
back to a class variable when the instance variable is missing. We
install data descriptors in the class dictionary, however, which
prevents taking advantage of this mechanism. Fully implementing the
cpython behavior would require a custom `__setattr__` on our class
types, which means using a metaclass for them.

So instead we implement a much more restricted form of class
variables.  Only class variables that are actually typed as `ClassVar`
are considered to be real class variables. They can be accessed
through the class. All other class variables are treated as simply
being default values for attributes. These default values can't be
accessed via the class, so they can't be changed.

When a non-Optional field is intialized to `None` in the class, we
simply ignore the default value. We could instead store an optional
type in the object and cast it away on attribute access, but this
would be extra work and increased code size with the only real benefit
being that access raises a `TypeError` instead of an `AttributeError`.

Closes mypyc/mypyc#218.
Closes mypyc/mypyc#186.