Support subclassing python classes (mypyc/mypyc#339)
I think we are going to want to support both options in mypyc/mypyc#296, and so this is
work on option 1 of it.
This is not fully finished code, but I wanted to ask for thoughts on it now.

My approach for this is moving class creation into genops, evaluating the base 
classes, calling a prim to create the class, and saving it with a newly added
`InitStatic` op.

In addition to supporting subclassing python classes during the migration,
we also want to be able to subclass `Generic[T]`. This turns out to be a huge
pain, because pre-3.7 `Generic` is implemented using metaclasses and 3.7
`Generic` is implemented using some other language features we didn't support
previously. We can't actually create out class using the metaclass constructor, 
so we allocate a *parallel* class with the metaclass and then copy the attributes 
over. This is pretty awful.