Allow overloaded overrides if they don't extend domain (#5505)
Fixes https://github.com/python/mypy/issues/4985

Mypy previously disallowed introducing overloads for operator methods in
overrides because this can interact unsafely with reverse methods. However it
looks like this is safe for the situations where the domain of the override
is not extended.

In fact the same unsafety exists for non-overloaded op-methods (see example
in https://github.com/python/mypy/issues/4985#issuecomment-414095444), but it
looks like we are fine with this, I have found few existing tests explicitly
testing for this.