Always add object to bases before calculating MRO (#5533)
Fixes #5527 

The problem is that MRO calculation algorithm may fail if certain corner cases if a base class list is empty (which is the case before class definition is analyzed). This can happen for example in an import cycle. The solution is to always add a temporary `object` base for un-analyzed classes, for them MRO will be anyway recomputed during third pass.

Ideally we would move MRO calculation to third pass, but this is hard, see https://github.com/python/mypy/issues/5536