Fix meet for tuple with instance (#5641)
Fixes https://github.com/python/mypy/issues/5640

Since generic tuple types are not supported, `map_instance_to_supertype` is not needed here. This solution doesn't work for general multiple inheritance, for example the test case will not work with `if isinstance(o, BaseTuple)`, but this is exactly the same as for e.g. two instance types.