Mark --quick-and-dirty as deprecated. (#5737)
Note that #5728 is not fixed by this yet! The behavior is unchanged
except it is no longer listed in the `--help` output.

Also add `file=sys.stderr` to a few of the `print()` calls for other
deprecated options.

Also mark it deprecated in the docs (but don't delete those docs yet).