Remove fscache arg from two functions in modulefinder.py (#5727)
In get_site_packages_dirs() the fscache argument was only used to
check for easy-install.pth, and that check appears to be done only
once.  And compute_search_paths() just passes it to
get_site_packages_dirs().  All this required a few places to create a
throwaway FileSystemCache instance.  It seems simpler to just os
os.path.isfile() and drop the fscache argument.