Make descriptor assignments more like normal assignments (#5853)
Fixes https://github.com/python/mypy/issues/5851

This makes descriptor assignments behave more like normal assignments. This includes:
* The same error message as for normal assignment
* Using binder to restrict types (but under a bit more strict condition)

I was not able to refactor all the logic to `checkmember.py`, but this is probably OK. Also this PR can introduce potential unsafety as discussed in the issue, but this is probably not much different from using binder in some other situations.