Make literal exprs have inferred type of 'Literal' based on context (#5990)
This pull request modifies the type checking logic so that literal
expressions will have an inferred type of 'Literal' if the context asks
for a literal type. That is, it implements support for this:

    x: Literal[1] = 1
    y = 1

    reveal_type(x)  # E: Revealed type is 'Literal[1]'
    reveal_type(y)  # E: Revealed type is 'builtins.int'

This pull requests also implements the `visit_literal_type` method
in the  `constraints.ConstraintBuilderVisitor` and `join.TypeJoinVisitor`
methods. Both visitors are exercised indirectly through the "let's use
literal types in collection contexts" code, but only the latter is
tested directly: I wasn't really sure how to directly test
`ConstraintBuilderVisitor`.

The implementation is simple though -- I'm pretty sure literal types
count as a "leaf type" so it's fine to return an empty list
(no constraints).